feat(server): include the eventmachine  on telemetryServer
Here we clean up a bit by including better context handling for shutdown.  We
also make use of an eventmachine attached to the telemetryServer so events can
be written directly without passing through the RPC.  Otherwise the server had
to include a client to the rpc that may yet be started by the time we rely on
it, which seems less good.